// src/api/model/contactStatsResponseDto.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * accounting
 * Accounting API description
 * OpenAPI spec version: 1.0
 */
import type { ContactInvoiceStatsDto } from './contactInvoiceStatsDto';

export interface ContactStatsResponseDto {
  issued: ContactInvoiceStatsDto;
  received: ContactInvoiceStatsDto;
}